+++ app/src/main/java/com/tosslab/jandi/app/events/entities/TopicInfoUpdateEvent.java
@@ -4,13 +4,13 @@ package com.tosslab.jandi.app.events.entities;
  * Created by Steve SeongUg Jung on 15. 4. 9..
  */
 public class TopicInfoUpdateEvent {
-    private final int id;
+    private final long id;
 
-    public TopicInfoUpdateEvent(int id) {
+    public TopicInfoUpdateEvent(long id) {
         this.id = id;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 }
