+++ app/src/main/java/com/tosslab/jandi/app/events/files/ConfirmDeleteFileEvent.java
@@ -5,13 +5,13 @@ package com.tosslab.jandi.app.events.files;
  * Created by Steve SeongUg Jung on 15. 2. 12..
  */
 public class ConfirmDeleteFileEvent {
-    private final int fileId;
+    private final long fileId;
 
-    public ConfirmDeleteFileEvent(int fileId) {
+    public ConfirmDeleteFileEvent(long fileId) {
         this.fileId = fileId;
     }
 
-    public int getFileId() {
+    public long getFileId() {
         return fileId;
     }
 }
