+++ app/src/main/java/com/tosslab/jandi/app/local/orm/repositories/ReadyMessageRepository.java
@@ -61,7 +61,7 @@ public class ReadyMessageRepository {
         return 0;
     }
 
-    public ReadyMessage getReadyMessage(int roomId) {
+    public ReadyMessage getReadyMessage(long roomId) {
         try {
             Dao<ReadyMessage, ?> readyMessageDao = helper.getDao(ReadyMessage.class);
             ReadyMessage readyMessage = readyMessageDao.queryBuilder()
