+++ app/src/main/java/com/tosslab/jandi/app/network/models/ReqDeleteTopic.java
@@ -4,7 +4,7 @@ package com.tosslab.jandi.app.network.models;
  * Created by justinygchoi on 2014. 5. 28..
  */
 public class ReqDeleteTopic {
-    public int teamId;
+    public long teamId;
 
     public ReqDeleteTopic(long teamId) {
         this.teamId = teamId;
