+++ app/src/main/java/com/tosslab/jandi/app/ui/filedetail/FileDetailPresenter.java
@@ -673,7 +673,7 @@ public class FileDetailPresenter {
 
         void showCheckNetworkDialog();
 
-        void showDeleteFileDialog(int fileId);
+        void showDeleteFileDialog(long fileId);
 
         void showUserInfoDialog(FormattedEntity user);
